move scenari_manager and experiemnt from namespace util to scenario
