+++ cmake/headers.cmake
@@ -64,14 +64,12 @@ include/warthog/search/vl_gridmap_expansion_policy.h
 include/warthog/util/cast.h
 include/warthog/util/cost_table.h
 include/warthog/util/dimacs_parser.h
-include/warthog/util/experiment.h
 include/warthog/util/file_utils.h
 include/warthog/util/gm_header.h
 include/warthog/util/helpers.h
 include/warthog/util/intrin.h
 include/warthog/util/macros.h
 include/warthog/util/pqueue.h
-include/warthog/util/scenario_manager.h
 include/warthog/util/string.h
 include/warthog/util/template.h
 include/warthog/util/timer.h
